Rename classes and modules for better abstraction
