+++ et_replay/comm/backend/pytorch_dist_backend.py
@@ -15,8 +15,10 @@ import torch.distributed as dist
 import torch.nn as nn
 
 from et_replay.comm.param_profile import paramProfile
-from et_replay.comm.pytorch_backend_utils import backendFunctions, collectiveArgsHolder
-
+from et_replay.comm.backend.base_backend import (
+    BaseBackend,
+    collectiveArgsHolder,
+)
 
 try:
     from param_bench.train.comms.pt.fb.internals import (
@@ -68,7 +70,7 @@ def _dequantize(obj):
             return resultTensor
 
 
-class PyTorchDistBackend(backendFunctions):
+class PyTorchDistBackend(BaseBackend):
     def get_collective_group(self, collectiveArgs):
         if self.use_ext_dist:
             return collectiveArgs.group.my_pg
