+++ et_replay/comm/commsTraceParser.py
@@ -7,7 +7,7 @@ from typing import List, Tuple
 from et_replay import ExecutionTrace
 from et_replay.comm import comms_utils
 from et_replay.comm.comms_utils import commsArgs
-from et_replay.comm.pytorch_backend_utils import supportedP2pOps
+from et_replay.comm.backend.base_backend import supportedP2pOps
 
 tensorDtypeMap = {
     "Tensor(int)": "int",
